AMP-17988 Pledges docuemnt size calculated incorrectly
git-svn-id: file:///svn/branches/AMP_2_10_RELEASE@24259 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/fundingpledges/action/SelectPledgeProgram.java
@@ -142,7 +142,7 @@ public class SelectPledgeProgram extends Action {
 			tdd.setName(file.getFileName());
 			tdd.setDescription("");
 			tdd.setNotes("");
-			tdd.setFileSize(file.getFileSize() / 1024 / 1024);
+			tdd.setFileSize((double)file.getFileSize() / 1024 / 1024);
 			tdd.setFormFile(file);
 			ActionMessages errors = new ActionMessages();
 			//NodeWrapper wrapper = tdd.saveToRepository(request, errors);
